requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
Django==3.2.18
django-appconf==1.0.5
django-ckeditor==6.5.1
django-extensions==3.2.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-nested-formset==0.1.4
django-taggit==4.0.0
django-taggit-templatetags2==1.6.1
executing==1.2.0
idna==3.4
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.15.1
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
